+++ R/pdynmc_furtherHelperFcts.R
@@ -931,6 +931,7 @@ dat.closedFormExpand.fct		<- function(
 
   varnames.temp	 <- c(varname.reg.estParam, varname.y)
 
+#  data.temp		<- dat.na[dat.na[, varname.i] == as.numeric(as.factor(i)), varnames.temp]
   data.temp		<- dat.na[dat.na[, varname.i] == as.numeric(i), varnames.temp]
 
   dat.temp		<- do.call(what = sub.clForm.fct, args = list(i = i, varname.i = varname.i, varname = varnames.temp, varname.y = varname.y
@@ -969,6 +970,7 @@ dat.expand.fct		<- function(
 
   varnames.temp	<- c(varname.reg.estParam, varname.y)
 
+#  data.temp		<- dat.na[dat.na[, varname.i] == as.numeric(as.factor(i)), varnames.temp]
   data.temp		<- dat.na[dat.na[, varname.i] == as.numeric(i), varnames.temp]
 
   dat.temp		<- data.temp[-c(1:max.lagTerms), ]