+++ inst/tests/arguments_for_StepByStepFunctionChecksEducData.R
@@ -7,60 +7,69 @@ rm(list = ls())
 library(pdynmc)
 
 
-data(ABdata, package = "pdynmc")
-dat <- ABdata
-dat[,c(4:7)] <- log(dat[,c(4:7)])
+data(cigDemand, package = "pdynmc")
+dat <- cigDemand
+
+
+dat$y			<- log(dat$packpc)				# response variable log(Q^Cigarettes)
+dat$x			<- log(dat$avgprs/dat$cpi)			# endogenous variable log(P^Cigarettes)
+dat$z1		<- (dat$taxs - dat$tax)/dat$cpi		# instrument 1 SalesTax
+dat$z2		<- dat$tax/dat$cpi				# instrument 2 CigTax
+dat$w			<- log((dat$income/dat$pop)/dat$cpi)	# exogenous variable log(rincome)
+
+
+
 
 
 
 dat = dat
-varname.i = "firm"
+varname.i = "state"
 varname.t = "year"
 
-use.mc.diff = TRUE
-use.mc.lev = FALSE
+use.mc.diff = FALSE
+use.mc.lev = TRUE
 use.mc.nonlin = FALSE
 use.mc.nonlinAS = NULL
 inst.stata = FALSE
 
-include.y = TRUE
-varname.y = "emp"
-lagTerms.y = 2
-maxLags.y = NULL
+include.y = FALSE
+varname.y = "y"
+lagTerms.y = 1
+maxLags.y = 4
 
-include.x = FALSE
+include.x = TRUE
 varname.reg.end = NULL
 lagTerms.reg.end = NULL
 maxLags.reg.end = NULL
 varname.reg.pre = NULL
 lagTerms.reg.pre = NULL
 maxLags.reg.pre = NULL
-varname.reg.ex = NULL
-lagTerms.reg.ex = NULL
-maxLags.reg.ex = NULL
-include.x.instr = FALSE
-varname.reg.instr = NULL
-inst.reg.ex.expand = TRUE
-include.x.toInstr = FALSE
-varname.reg.toInstr = NULL
+varname.reg.ex = c("z1")
+lagTerms.reg.ex = c(0)
+maxLags.reg.ex = c(2)
+include.x.instr = TRUE
+varname.reg.instr = c("z1")
+inst.reg.ex.expand = FALSE
+include.x.toInstr = TRUE
+varname.reg.toInstr = "x"
 
 fur.con = TRUE
 fur.con.diff = TRUE
 fur.con.lev = FALSE
-varname.reg.fur = c("wage", "capital", "output")
-lagTerms.reg.fur = c(1,2,2)
+varname.reg.fur = "w"
+lagTerms.reg.fur = 0
 
-include.dum = TRUE
-dum.diff = TRUE
-dum.lev = FALSE
-varname.dum = "year"
+include.dum = FALSE
+dum.diff = NULL
+dum.lev = NULL
+varname.dum = NULL
 
 col_tol = 0.65
 w.mat = "iid.err"
 w.mat.stata = FALSE
 
 std.err = "corrected"
-estimation = "onestep"
+estimation = "twostep"
 max.iter = 4
 iter.tol = 0.01
 inst.thresh = NULL