+++ NEWS.md
@@ -1,6 +1,37 @@
+# pdynmc version 0.9.7
+
+Update of version 0.9.6 which updates the estimation function, the functions for visualizing the panel data structure, and adds two datasets to the package. The functionality for deriving instruments and estimating parameters: Covariates for which no parameters are estimated, but from which instruments are derived and covariates for which parameters for which parameters are estimated, but from which no instruments are derived.
+
+
+### pdynmc
+* Add flexibility for covariates for which no parameters are estimated, but which are used as instruments (function arguments: `include.x.instr`, `varname.reg.instr`)
+* Add flexibility for covariates for which parameters are estimated, but which are not used as instruments (function arguments: `include.x.toInstr`, `varname.reg.toInstr`)
+* Adjust code for further data structures
+* Update function documentation
+* Fix bug in function that creates instrument matrix
+
+
+### data.info
+* Adjust code for further data structures
+
+
+### struc.UPD.plot
+* Adjust code for further data structures
+
+
+
+### datasets
+* Add dataset of Arellano and Bond (1991) on employment by firms located in the UK
+* Add dataset of Stock and Watson (2003) on cigarette consumption in the US
+
+
+
+
+
 # pdynmc version 0.9.6
 
 Minor update of version 0.9.5 which adds doubly corrected standard errors. Also, commits and suggestions of github user tappek are added. Additionally, the compatibility of the estimation function with further input data structures is improved and a bug in the estimation function when multiple instruments from non-lagged dependent endogenous covariates are derived is corrected.
+
 
 ### pdynmc
 * Implement doubly corrected standard errors
@@ -8,12 +39,15 @@ Minor update of version 0.9.5 which adds doubly corrected standard errors. Also,
 * Ensure compatibility with `tibble` data frames
 * Correct bug in deriving instruments from multiple endogenous covariates
 
+
 ### wald.fct
 * Adjust structure of function according to S3 class htest.object
+
 
 ### jtest.fct
 * Adjust structure of function according to S3 class htest.object
 * Minor adjustment to avoid partial argument matching
+
 
 ### mtest.fct
 * Adjust structure of function according to S3 class htest.object
@@ -28,6 +62,7 @@ Minor update of version 0.9.5 which adds doubly corrected standard errors. Also,
 
 Minor update of version 0.9.4 which adds further functionality and argument checks to estimation function. Additionally, the computation underlying non-robust two-step standard errors is adjusted (option accessible by changing argument "std.err" from its default to "std.err = unadjusted") and the functions for deriving instruments from further exogenous covariates were adjusted to comply with data requirements.
 
+
 ### pdynmc
 * Allow lagged dependent variable to be instrumented
 * Extend checks of function arguments
@@ -41,6 +76,7 @@ Minor update of version 0.9.4 which adds further functionality and argument chec
 # pdynmc version 0.9.4
 
 Minor update of version 0.9.3 in which package DESCRIPTION, CITATION, and documentation is adjusted and two further package vignettes are added.
+
 
 ### new vignettes:
 *pdynmc-intro